# Limit Forward references in Mathtext parser