… lib/matplotlib/mathtext.pyi