## PR summary

Comment indicates that limiting forward references is important for
performance, but many more were used than needed.

This is a 75% reduction in forward reference instances.

Many of these are included in the definition of remaining forward
references but do not themselves need to be a forward reference (indeed
only 1 thing in a given cycle actually needs to be such that it can be
referred to prior to its definition)

`auto_delim` is recursively self-referential, and thus _must_ be all on
its own

`placeable` and `accent` form a tight loop, so one of those two needed
to be a forward reference, and placeable was the easiest to keep (and
most likely to be used in another definition that may introduce similar
constraints)

`token` is one of the more generic definitions, that is used in multiple
other definitions.

`required_group` and `optional_group` both caused failed tests when I
tried to make them no longer forward references, has something to do
with how `__call__` works for `Forward` vs `And` objects in pyparsing
(though did not dig too much deeper into it, beyond noticing that the
names in future `required_group("name")` calls still had error messages
that say `"group"`, not their newer names, and that reverting to
`Forward` instances fixed it).

`operatorname` needed to move to avoid using forward references for
`simple`, but does not actually have a cycle.

Otherwise rewrapped some of the definitions to be one line and fixed a
typing error that I noticed when looking into the parsing things.


## PR checklist
<!-- Please mark any checkboxes that do not apply to this PR as [N/A].-->

- [N/A] "closes #0000" is in the body of the PR description to [link the related issue](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
- [N/A] new and changed code is [tested](https://matplotlib.org/devdocs/devel/testing.html)
- [N/A] *Plotting related* features are demonstrated in an [example](https://matplotlib.org/devdocs/devel/documenting_mpl.html#writing-examples-and-tutorials)
- [N/A] *New Features* and *API Changes* are noted with a [directive and release note](https://matplotlib.org/devdocs/devel/coding_guide.html#new-features-and-api-changes)
- [N/A] Documentation complies with [general](https://matplotlib.org/devdocs/devel/documenting_mpl.html#writing-rest-pages) and [docstring](https://matplotlib.org/devdocs/devel/documenting_mpl.html#writing-docstrings) guidelines


@devRD I would appreciate your review on this. If this reorg would cause conflicts for other PRs already in flight, it can absolutely wait.
I just saw the comment saying to minimize usage of Forward and then was immediately like "why is this a Forward" and kept going.

The changes themselves are relatively minimal and would be easy to redo if needed:

- delete line that instantiates the Forward
- delete the `<<` from `<<=` in the redefinition 
- a few reorderings to maintain directionality without Forword references
- keep anything that is referred to before it is defined (and cannot be reordered) as a Forward  (plus the optional/required_group)

<!--
Thank you so much for your PR!  To help us review your contribution, please
consider the following points:

- A development guide is available at https://matplotlib.org/devdocs/devel/index.html.

- Help with git and github is available at https://matplotlib.org/devdocs/devel/development_workflow.html

- Create a separate branch for your changes and open the PR from this branch. Please avoid working on `main`.

- The PR title should summarize the changes, for example "Raise ValueError on
  non-numeric input to set_xlim".  Avoid non-descriptive titles such as
  "Addresses issue #8576".

- The summary should provide at least 1-2 sentences describing the pull request
  in detail (Why is this change required?  What problem does it solve?) and
  link to any relevant issues.

- If you are contributing fixes to docstrings, please pay attention to
  https://matplotlib.org/stable/devel/documenting_mpl.html#formatting-conventions.  In particular,
  note the difference between using single backquotes, double backquotes, and
  asterisks in the markup.

We understand that PRs can sometimes be overwhelming, especially as the
reviews start coming in.  Please let us know if the reviews are unclear or
the recommended next step seems overly demanding, if you would like help in
addressing a reviewer's comments, or if you have been waiting too long to hear
back on your PR.
-->