@ksunden 2023-06-27
As for performance claims, it (in not that high of N) seems to be about a 5% speed up, judging from time the time it takes to run `pytest test_mathtext.py` (26.5 seconds compared to 28.1 seconds on main)(which has some significant overhead, but may be a bit more realistic on what even heavy users of mathtext may see) 

Running on `[parser.parse(a) for a in math_tests]` (where `math_tests` is the list from the mathtext tests, stripped of any Nones. It has length of 77, so memoising should be minimized which is good for testing speed of parsing:

main:

`387 ms ± 2.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

on pr branch:

`301 ms ± 1.43 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

So that is a 22% improvement focusing on  parsing, which is honestly more than I thought it would be (granted, parsing mathtext is rarely the bottleneck, but still).