@devRD 2023-06-27
I think the changes are really helpful!
Not sure how many PRs it affects currently, but with these new changes I think one will have to update the formatting for any new commands they have added, which is a minimal change. I believe we are good to go ahead with this.